requirements.txt
PyQt6
google-generativeai
edge-tts
pygame 
speechrecognition 
asyncio 
plocate